requirements/base.txt
numpy>=1.24
PyYAML>=6.0.1
pydantic>=2.6
python-dotenv>=1.0.1
requests>=2.31
opencv-python-headless>=4.8
flask>=3.0
waitress>=2.1
